[omp] Making OpenMP work under newest MacOS

+++ include/mitsuba/core/thread.h
@@ -199,10 +199,10 @@ private:
 #if defined(MTS_OPENMP)
 #if defined(__OSX__)
 /// Variant of \c omp_get_max_threads that works on OSX
-extern MTS_EXPORT_CORE int mts_omp_get_max_threads();
+extern MTS_EXPORT_CORE size_t mts_omp_get_max_threads();
 
 /// Variant of \c omp_get_thread_num that works on OSX
-extern MTS_EXPORT_CORE int mts_omp_get_thread_num();
+extern MTS_EXPORT_CORE size_t mts_omp_get_thread_num();
 #else
 #define mts_omp_get_max_threads omp_get_max_threads
 #define mts_omp_get_thread_num omp_get_thread_num
